consolidate FilePathTests into a single location
