+++ src/cpp/shared_core/FilePathTests.cpp
@@ -27,6 +27,44 @@
 
 namespace rstudio {
 namespace core {
+namespace {
+
+#ifdef _WIN32
+
+// helper for creating a path with the system drive
+// prefixed for Windows. we try to use the current
+// drive if at all possible; if we cannot retrieve
+// that for some reason then we just fall back to
+// the default system drive (normally C:)
+std::string getDrivePrefix()
+{
+   char buffer[MAX_PATH];
+   DWORD n = GetCurrentDirectory(MAX_PATH, buffer);
+   if (n < 2)
+      return ::getenv("SYSTEMDRIVE");
+
+   if (buffer[1] != ':')
+      return ::getenv("SYSTEMDRIVE");
+
+   return std::string(buffer, 2);
+}
+
+FilePath createPath(const std::string& path = "/")
+{
+   static const std::string prefix = getDrivePrefix();
+   return FilePath(prefix + path);
+}
+
+#else
+
+FilePath createPath(const std::string& path = "/")
+{
+   return FilePath(path);
+}
+
+#endif /* _WIN32 */
+
+} // end anonymous namespace
 
 TEST_CASE("Empty File Path tests")
 {
@@ -34,6 +72,20 @@ TEST_CASE("Empty File Path tests")
    {
       FilePath f;
       CHECK(f.getAbsolutePath().empty());
+   }
+
+   SECTION("relative path construction")
+   {
+      FilePath rootPath = createPath();
+      FilePath pPath = createPath("/path/to");
+      FilePath aPath = createPath("/path/to/a");
+      FilePath bPath = createPath("/path/to/b");
+
+      CHECK(aPath.isWithin(pPath));
+      CHECK(bPath.isWithin(pPath));
+      CHECK(!aPath.isWithin(bPath));
+
+      CHECK(aPath.getRelativePath(pPath) == "a");
    }
 
    SECTION("Comparison (equal, true)")
@@ -129,6 +181,89 @@ TEST_CASE("Empty File Path tests")
       CHECK(f1.completeChildPath("/some/absolute/path", f2)); // Error here.
       CHECK(f2 == f1); // f2 should have been set to f1.
    }
+
+   SECTION("child path completion")
+   {
+      // simple path completion should do what's expected
+      FilePath aPath = createPath("/path/to/a");
+      FilePath bPath = createPath("/path/to/a/b");
+      CHECK(aPath.completeChildPath("b") == bPath);
+
+      // trying to complete to a path outside should fail and return the original path
+      FilePath cPath = createPath("/path/to/foo");
+      CHECK(cPath.completeChildPath("../bar") == cPath);
+      CHECK(cPath.completeChildPath("/path/to/quux") == cPath);
+
+      // trailing slashes are okay
+      FilePath dPath = createPath("/path/to/");
+      FilePath ePath = createPath("/path/to/e");
+      CHECK(dPath.completeChildPath("e") == ePath);
+   }
+
+   SECTION("general path completion")
+   {
+      // simple path completion should do what's expected
+      FilePath aPath = createPath("/path/to/a");
+      FilePath bPath = createPath("/path/to/a/b");
+      CHECK(aPath.completePath("b") == bPath);
+
+      // absolute paths are allowed
+      FilePath cPath = createPath("/path/to/c");
+      FilePath dPath = createPath("/path/to/d");
+      CHECK(cPath.completePath("/path/to/d") == dPath);
+
+      // directory traversal is allowed
+      FilePath ePath = createPath("/path/to/e");
+      FilePath fPath = createPath("/path/to/f");
+      CHECK(ePath.completePath("../f").getLexicallyNormalPath() == fPath.getAbsolutePath());
+   }
+
+#ifdef _WIN32
+
+   SECTION("relative paths for UNC shares")
+   {
+      // NOTE: need to be robust against mixed separators as these can
+      // leak in depending on the API used to request the file path.
+      //
+      // https://github.com/rstudio/rstudio/issues/6587
+      FilePath pPath(R"(//LOCALHOST/c$/p)");
+      FilePath aPath(R"(\\LOCALHOST\c$\p\a)");
+      CHECK(aPath.getRelativePath(pPath) == "a");
+   }
+
+#else
+
+   // Non-Windows tests
+
+   SECTION("directory write testing")
+   {
+      // create temporary directory
+      FilePath tempDir;
+      Error error = FilePath::tempFilePath(tempDir);
+      CHECK(!error);
+
+      error = tempDir.ensureDirectory();
+      CHECK(!error);
+
+      // ensure it's writeable
+      error = tempDir.changeFileMode(FileMode::USER_READ_WRITE_ALL_READ);
+      CHECK(!error);
+
+      // it should now report as writeable
+      bool writeable = false;
+      error = tempDir.isWriteable(writeable);
+      CHECK(!error);
+      CHECK(writeable);
+
+      // however, it should be an error to test write permissions (that's only for files)
+      error = tempDir.testWritePermissions();
+      CHECK(error);
+
+      // clean up
+      tempDir.remove();
+   }
+#endif /* !_WIN32 */
+
 }
 
 TEST_CASE("Copy FilePath Tests")
